[#10026] Make TraceService, SqlMetaDataService abstract
