+++ collector/src/main/java/com/navercorp/pinpoint/collector/handler/grpc/GrpcSpanChunkHandler.java
@@ -6,6 +6,7 @@ import com.navercorp.pinpoint.collector.handler.SimpleHandler;
 import com.navercorp.pinpoint.collector.service.TraceService;
 import com.navercorp.pinpoint.common.server.bo.SpanChunkBo;
 import com.navercorp.pinpoint.common.server.bo.grpc.GrpcSpanFactory;
+import com.navercorp.pinpoint.common.util.CollectionUtils;
 import com.navercorp.pinpoint.grpc.Header;
 import com.navercorp.pinpoint.grpc.MessageFormatUtils;
 import com.navercorp.pinpoint.grpc.server.ServerContext;
@@ -14,10 +15,11 @@ import com.navercorp.pinpoint.grpc.trace.PSpanEvent;
 import com.navercorp.pinpoint.grpc.trace.PTransactionId;
 import com.navercorp.pinpoint.io.request.ServerRequest;
 import io.grpc.Status;
-import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.stereotype.Service;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -30,13 +32,15 @@ public class GrpcSpanChunkHandler implements SimpleHandler<GeneratedMessageV3> {
     private final Logger logger = LogManager.getLogger(getClass());
     private final boolean isDebug = logger.isDebugEnabled();
 
-    private final TraceService traceService;
+    private final TraceService[] traceServices;
 
     private final GrpcSpanFactory spanFactory;
 
-    public GrpcSpanChunkHandler(TraceService traceService, GrpcSpanFactory spanFactory) {
-        this.traceService = Objects.requireNonNull(traceService, "traceService");
+    public GrpcSpanChunkHandler(TraceService[] traceServices, GrpcSpanFactory spanFactory) {
+        this.traceServices = Objects.requireNonNull(traceServices, "traceServices");
         this.spanFactory = Objects.requireNonNull(spanFactory, "spanFactory");
+
+        logger.info("TraceServices {}", Arrays.toString(traceServices));
     }
 
     @Override
@@ -56,12 +60,15 @@ public class GrpcSpanChunkHandler implements SimpleHandler<GeneratedMessageV3> {
             logger.debug("Handle PSpanChunk={}", createSimpleSpanChunkLog(spanChunk));
         }
 
-        try {
-            final Header agentInfo = ServerContext.getAgentInfo();
-            final SpanChunkBo spanChunkBo = spanFactory.buildSpanChunkBo(spanChunk, agentInfo);
-            this.traceService.insertSpanChunk(spanChunkBo);
-        } catch (Exception e) {
-            logger.warn("Failed to handle spanChunk={}", MessageFormatUtils.debugLog(spanChunk), e);
+
+        final Header agentInfo = ServerContext.getAgentInfo();
+        final SpanChunkBo spanChunkBo = spanFactory.buildSpanChunkBo(spanChunk, agentInfo);
+        for (TraceService traceService : traceServices) {
+            try {
+                traceService.insertSpanChunk(spanChunkBo);
+            } catch (Exception e) {
+                logger.warn("Failed to handle spanChunk={}", MessageFormatUtils.debugLog(spanChunk), e);
+            }
         }
     }
 
@@ -70,7 +77,7 @@ public class GrpcSpanChunkHandler implements SimpleHandler<GeneratedMessageV3> {
             return "";
         }
 
-        StringBuilder log = new StringBuilder();
+        StringBuilder log = new StringBuilder(64);
 
         PTransactionId transactionId = spanChunk.getTransactionId();
         log.append(" transactionId:");
@@ -78,17 +85,16 @@ public class GrpcSpanChunkHandler implements SimpleHandler<GeneratedMessageV3> {
 
         log.append(" spanId:").append(spanChunk.getSpanId());
 
-
-        StringBuilder spanEventSequenceLog = new StringBuilder();
-        List<PSpanEvent> spanEventList = spanChunk.getSpanEventList();
-        for (PSpanEvent pSpanEvent : spanEventList) {
-            if (pSpanEvent == null) {
-                continue;
+        final List<PSpanEvent> spanEventList = spanChunk.getSpanEventList();
+        if (CollectionUtils.hasLength(spanEventList)) {
+            log.append(" spanEventSequence:");
+            for (PSpanEvent pSpanEvent : spanEventList) {
+                if (pSpanEvent == null) {
+                    continue;
+                }
+                log.append(pSpanEvent.getSequence()).append(" ");
             }
-            spanEventSequenceLog.append(pSpanEvent.getSequence()).append(" ");
         }
-
-        log.append(" spanEventSequence:").append(spanEventSequenceLog.toString());
 
         return log.toString();
     }
