+++ collector/src/main/java/com/navercorp/pinpoint/collector/service/HbaseSqlMetaDataService.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2018 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.collector.service;
+
+import com.navercorp.pinpoint.collector.dao.SqlMetaDataDao;
+import com.navercorp.pinpoint.common.server.bo.SqlMetaDataBo;
+import org.springframework.stereotype.Service;
+
+import java.util.Objects;
+
+@Service
+public class HbaseSqlMetaDataService implements SqlMetaDataService {
+    private final SqlMetaDataDao sqlMetaDataDao;
+
+    public HbaseSqlMetaDataService(SqlMetaDataDao sqlMetaDataDao) {
+        this.sqlMetaDataDao = Objects.requireNonNull(sqlMetaDataDao, "sqlMetaDataDao");
+    }
+
+    @Override
+    public void insert(final SqlMetaDataBo sqlMetaDataBo) {
+        this.sqlMetaDataDao.insert(sqlMetaDataBo);
+    }
+}
